Updated sorting interface and implementation.

+++ src/main/java/ro/jademy/carrentalshop/model/cars/dacia/Dacia.java
@@ -9,7 +9,7 @@ public class Dacia extends Car {
 
     }
 
-    public Dacia(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, double pricePerDay, String carType) {
+    public Dacia(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, long pricePerDay, String carType) {
         super("Dacia", model, gearbox, fuelType, noOfDoors, noOfSeats, pricePerDay, carType);
     }
 }
