+++ src/main/java/ro/jademy/carrentalshop/model/cars/dacia/DaciaLodgy.java
@@ -6,7 +6,7 @@ public class DaciaLodgy extends Dacia {
     public DaciaLodgy() {
     }
 
-    public DaciaLodgy(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, double pricePerDay, String carType) {
+    public DaciaLodgy(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, long pricePerDay, String carType) {
         super(model, gearbox, fuelType, noOfDoors, noOfSeats, pricePerDay, carType);
     }
 }
