+++ src/main/java/ro/jademy/carrentalshop/model/cars/dacia/DaciaLogan.java
@@ -5,7 +5,7 @@ public class DaciaLogan extends Dacia {
     public DaciaLogan() {
     }
 
-    public DaciaLogan(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, double pricePerDay, String carType) {
+    public DaciaLogan(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, long pricePerDay, String carType) {
         super(model, gearbox, fuelType, noOfDoors, noOfSeats, pricePerDay, carType);
     }
 }
