+++ src/main/java/ro/jademy/carrentalshop/model/cars/ford/Ford.java
@@ -9,7 +9,7 @@ public class Ford extends Car {
 
     }
 
-    public Ford(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, double pricePerDay, String carType) {
+    public Ford(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, long pricePerDay, String carType) {
         super("Ford", model, gearbox, fuelType, noOfDoors, noOfSeats, pricePerDay, carType);
     }
 }
