+++ src/main/java/ro/jademy/carrentalshop/model/cars/ford/FordKuga.java
@@ -4,7 +4,7 @@ public class FordKuga extends Ford {
     public FordKuga() {
     }
 
-    public FordKuga(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, double pricePerDay, String carType) {
+    public FordKuga(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, long pricePerDay, String carType) {
         super(model, gearbox, fuelType, noOfDoors, noOfSeats, pricePerDay, carType);
     }
 }
