+++ src/main/java/ro/jademy/carrentalshop/model/cars/mercedes/MercedesBenzCClass.java
@@ -4,7 +4,7 @@ public class MercedesBenzCClass extends MercedesBenz {
     public MercedesBenzCClass() {
     }
 
-    public MercedesBenzCClass(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, double pricePerDay, String carType) {
+    public MercedesBenzCClass(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, long pricePerDay, String carType) {
         super("Mercedes-Benz", model, gearbox, fuelType, noOfDoors, noOfSeats, pricePerDay, carType);
     }
 }
