+++ src/main/java/ro/jademy/carrentalshop/model/cars/mercedes/MercedesBenzGLC.java
@@ -4,7 +4,7 @@ public class MercedesBenzGLC extends MercedesBenz {
     public MercedesBenzGLC() {
     }
 
-    public MercedesBenzGLC(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, double pricePerDay, String carType) {
+    public MercedesBenzGLC(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, long pricePerDay, String carType) {
         super("Mercedes-Benz", model, gearbox, fuelType, noOfDoors, noOfSeats, pricePerDay, carType);
     }
 }
