+++ src/main/java/ro/jademy/carrentalshop/model/cars/opel/Opel.java
@@ -9,7 +9,7 @@ public class Opel extends Car {
 
     }
 
-    public Opel(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, double pricePerDay, String carType) {
+    public Opel(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, long pricePerDay, String carType) {
         super("Opel", model, gearbox, fuelType, noOfDoors, noOfSeats, pricePerDay, carType);
     }
 }
