+++ src/main/java/ro/jademy/carrentalshop/model/cars/opel/OpelCorsa.java
@@ -4,7 +4,7 @@ public class OpelCorsa extends Opel {
     public OpelCorsa() {
     }
 
-    public OpelCorsa(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, double pricePerDay, String carType) {
+    public OpelCorsa(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, long pricePerDay, String carType) {
         super(model, gearbox, fuelType, noOfDoors, noOfSeats, pricePerDay, carType);
     }
 }
