+++ src/main/java/ro/jademy/carrentalshop/model/cars/opel/OpelGrandland.java
@@ -4,7 +4,7 @@ public class OpelGrandland extends Opel {
     public OpelGrandland() {
     }
 
-    public OpelGrandland(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, double pricePerDay, String carType) {
+    public OpelGrandland(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, long pricePerDay, String carType) {
         super(model, gearbox, fuelType, noOfDoors, noOfSeats, pricePerDay, carType);
     }
 }
