+++ src/main/java/ro/jademy/carrentalshop/model/cars/opel/OpelVivaro.java
@@ -4,7 +4,7 @@ public class OpelVivaro extends Opel {
     public OpelVivaro() {
     }
 
-    public OpelVivaro(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, double pricePerDay, String carType) {
+    public OpelVivaro(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, long pricePerDay, String carType) {
         super(model, gearbox, fuelType, noOfDoors, noOfSeats, pricePerDay, carType);
     }
 }
