+++ src/main/java/ro/jademy/carrentalshop/model/cars/renault/RenaultCaptur.java
@@ -4,7 +4,7 @@ public class RenaultCaptur extends Renault {
     public RenaultCaptur() {
     }
 
-    public RenaultCaptur(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, double pricePerDay, String carType) {
+    public RenaultCaptur(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, long pricePerDay, String carType) {
         super(model, gearbox, fuelType, noOfDoors, noOfSeats, pricePerDay, carType);
     }
 }
