+++ src/main/java/ro/jademy/carrentalshop/model/cars/renault/RenaultMegane.java
@@ -4,7 +4,7 @@ public class RenaultMegane extends Renault {
     public RenaultMegane() {
     }
 
-    public RenaultMegane(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, double pricePerDay, String carType) {
+    public RenaultMegane(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, long pricePerDay, String carType) {
         super(model, gearbox, fuelType, noOfDoors, noOfSeats, pricePerDay, carType);
     }
 }
