+++ src/main/java/ro/jademy/carrentalshop/model/cars/renault/RenaultZoe.java
@@ -4,7 +4,7 @@ public class RenaultZoe extends Renault {
     public RenaultZoe() {
     }
 
-    public RenaultZoe(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, double pricePerDay, String carType) {
+    public RenaultZoe(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, long pricePerDay, String carType) {
         super(model, gearbox, fuelType, noOfDoors, noOfSeats, pricePerDay, carType);
     }
 }
