+++ src/main/java/ro/jademy/carrentalshop/model/cars/toyota/ToyotaAvensis.java
@@ -4,7 +4,7 @@ public class ToyotaAvensis extends Toyota {
     public ToyotaAvensis() {
     }
 
-    public ToyotaAvensis(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, double pricePerDay, String carType) {
+    public ToyotaAvensis(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, long pricePerDay, String carType) {
         super(model, gearbox, fuelType, noOfDoors, noOfSeats, pricePerDay, carType);
     }
 }
