+++ src/main/java/ro/jademy/carrentalshop/model/cars/toyota/ToyotaAygo.java
@@ -5,7 +5,7 @@ public class ToyotaAygo extends Toyota {
 
     }
 
-    public ToyotaAygo(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, double pricePerDay, String carType) {
+    public ToyotaAygo(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, long pricePerDay, String carType) {
         super(model, gearbox, fuelType, noOfDoors, noOfSeats, pricePerDay, carType);
     }
 }
