+++ src/main/java/ro/jademy/carrentalshop/model/cars/toyota/ToyotaCorolla.java
@@ -4,7 +4,7 @@ public class ToyotaCorolla extends Toyota {
     public ToyotaCorolla() {
     }
 
-    public ToyotaCorolla(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, double pricePerDay, String carType) {
+    public ToyotaCorolla(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, long pricePerDay, String carType) {
         super(model, gearbox, fuelType, noOfDoors, noOfSeats, pricePerDay, carType);
     }
 }
