+++ src/main/java/ro/jademy/carrentalshop/model/cars/toyota/ToyotaYaris.java
@@ -4,7 +4,7 @@ public class ToyotaYaris extends Toyota {
     public ToyotaYaris() {
     }
 
-    public ToyotaYaris(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, double pricePerDay, String carType) {
+    public ToyotaYaris(String model, String gearbox, String fuelType, int noOfDoors, int noOfSeats, long pricePerDay, String carType) {
         super(model, gearbox, fuelType, noOfDoors, noOfSeats, pricePerDay, carType);
     }
 }
